app/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
boto3==1.35.40
botocore==1.35.40
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
ecdsa==0.19.0
fastapi==0.115.2
greenlet==3.1.1
h11==0.14.0
httptools==0.6.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.1
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic[email]
pydantic_core==2.23.4
pymssql==2.3.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
PyYAML==6.0.2
rsa==4.9
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.40.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
watchfiles==0.24.0
websockets==13.1
